+++ react-components/src/components/RouteToMarket/index.jsx
@@ -2,88 +2,78 @@ import React, { useState, memo } from 'react'
 import PropTypes from 'prop-types'
 
 import { RouteToMarketSection } from '@src/components/RouteToMarket/RouteToMarketSection'
-import Services from '@src/Services'
-import { analytics, objectHasValue } from '@src/Helpers'
+import { objectHasValue } from '@src/Helpers'
 import { useDebounce } from '@src/components/hooks/useDebounce'
 import { AddButton } from '@src/components/ObjectivesList/AddButton/AddButton'
+import { useUpdate } from '@src/components/hooks/useUpdate/useUpdate'
 
-export const RouteToMarket = memo(({ fields, formData, formFields }) => {
-  const [routes, setRoutes] = useState(fields)
-  const [pushedAnalytic, setPushedAnalytic] = useState(false)
+export const RouteToMarket = memo(
+  ({ fields, formData, formFields, model_name }) => {
+    const [routes, setRoutes] = useState(fields)
 
-  const { companyexportplan, pk, ...lastField } = routes.length
-    ? routes[routes.length - 1]
-    : {}
+    const { companyexportplan, pk, ...lastField } = routes.length
+      ? routes[routes.length - 1]
+      : {}
+    const [update, create, deleteItem] = useUpdate('route-to-market')
 
-  const addTable = () => {
-    Services.createRouteToMarket({ ...formFields })
-      .then((data) => setRoutes([...routes, data]))
-      .then(() => {
-        const newElement = document.getElementById(
-          `Route to market ${routes.length + 1}`
-        ).parentNode
-        newElement.scrollIntoView()
-      })
-      .catch(() => {})
-  }
+    const addTable = () => {
+      create({ ...formFields })
+        .then((data) => setRoutes([...routes, data]))
+        .then(() => {
+          const newElement = document.getElementById(
+            `Route to market ${routes.length + 1}`
+          )
+          if (newElement) {
+            newElement.parentNode.scrollIntoView()
+          }
+        })
+    }
 
-  const deleteTable = (id) => {
-    Services.deleteRouteToMarket(id)
-      .then(() => {
+    const deleteTable = (id) => {
+      deleteItem({ pk: id, model_name }).then(() => {
         setRoutes(routes.filter((x) => x.pk !== id))
       })
-      .catch(() => {})
-  }
+    }
 
-  const update = (field, selected) => {
-    Services.updateRouteToMarket({ ...field, ...selected })
-      .then(() => {
-        if (!pushedAnalytic) {
-          analytics({
-            event: 'planSectionSaved',
-            sectionTitle: 'route-to-market',
-          })
-          setPushedAnalytic(true)
-        }
-      })
-      .catch(() => {})
-  }
+    const request = (field, selected) =>
+      update({ ...field, ...selected, model_name })
 
-  const debounceUpdate = useDebounce(update)
+    const debounceUpdate = useDebounce(request)
 
-  const onChange = (id, selected) => {
-    const field = routes.find((x) => x.pk === id)
-    const updatedRoutes = routes.map((x) =>
-      x.pk === id ? { ...x, ...selected } : x
+    const onChange = (id, selected) => {
+      const field = routes.find((x) => x.pk === id)
+      const updatedRoutes = routes.map((x) =>
+        x.pk === id ? { ...x, ...selected } : x
+      )
+
+      setRoutes(updatedRoutes)
+      debounceUpdate(field, selected)
+    }
+
+    return (
+      <>
+        {routes.length >= 1 &&
+          routes.map((field, id) => (
+            <RouteToMarketSection
+              key={field.pk}
+              {...formData}
+              data={formData.data.map((x) =>
+                x.name === 'route' ? { ...x, label: `${x.label} ${id + 1}` } : x
+              )}
+              field={field}
+              onChange={onChange}
+              deleteTable={deleteTable}
+            />
+          ))}
+        <AddButton
+          add={addTable}
+          isDisabled={routes.length ? !objectHasValue(lastField) : false}
+          cta="Add route to market"
+        />
+      </>
     )
-
-    setRoutes(updatedRoutes)
-    debounceUpdate(field, selected)
   }
-
-  return (
-    <>
-      {routes.length >= 1 &&
-        routes.map((field, id) => (
-          <RouteToMarketSection
-            key={field.pk}
-            {...formData}
-            data={formData.data.map((x) =>
-              x.name === 'route' ? { ...x, label: `${x.label} ${id + 1}` } : x
-            )}
-            field={field}
-            onChange={onChange}
-            deleteTable={deleteTable}
-          />
-        ))}
-      <AddButton
-        add={addTable}
-        isDisabled={routes.length ? !objectHasValue(lastField) : false}
-        cta="Add route to market"
-      />
-    </>
-  )
-})
+)
 
 RouteToMarket.propTypes = {
   fields: PropTypes.arrayOf(
@@ -111,4 +101,5 @@ RouteToMarket.propTypes = {
   formFields: PropTypes.objectOf(
     PropTypes.oneOfType([PropTypes.string, PropTypes.number])
   ).isRequired,
+  model_name: PropTypes.string.isRequired,
 }