+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 - GP2-1903 - New sign up page design implemented
 - GP2-1709 - trade barrier integration
 - GP2-2302 - ep dropdown validation
+- GP2-2336 - collapse api object
 
 ### Fixed bugs
 - GP2-2292 - Save product on name change only